requirements.txt
fastapi
sqlalchemy
requests
psycopg2-binary
uvicorn
python-dotenv
asyncpg
greenlet
httpx
coverage
pytest
pytest-asyncio
